+++ praw/__init__.py
@@ -31,12 +31,12 @@ import re
 import requests
 import six
 import sys
-from praw import decorators, errors
-from praw.handlers import DefaultHandler
-from praw.helpers import normalize_url
-from praw.internal import (_prepare_request, _raise_redirect_exceptions,
+import decorators, errors
+from handlers import DefaultHandler
+from helpers import normalize_url
+from internal import (_prepare_request, _raise_redirect_exceptions,
                            _raise_response_exceptions)
-from praw.settings import CONFIG
+from settings import CONFIG
 from requests.compat import urljoin
 from requests import Request
 from six.moves import html_entities, http_cookiejar